Update indicator and settings

+++ src/Trading.Bot/Services/TradeManager.cs
@@ -90,7 +90,7 @@ public class TradeManager : BackgroundService
 
         var generalTrend = longerTimeFrameCandles.CalcTrend(settings.Integers[0]).Last();
 
-        return generalTrend == (int)signal;
+        return generalTrend == signal;
     }
 
     private static bool GoodTradingTime()
